# backend/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Blog service",
    "description": "Blog API",
    "license": {
      "name": "MIT"
    }
  },
  "basePath": "/v1",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "summary": "get user",
        "operationId": "getUser",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "patch": {
        "summary": "update user",
        "operationId": "updateUser",
        "consumes": [
          "application/vnd.api+json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User patch document",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "required": [
                "id"
              ],
              "properties": {
                "id": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "$ref": "./definitions/User.json"
    }
  }
}